# openapi.yaml
openapi: 3.0.0
info:
  title: 'GeoService API'
  version: 2.2.0
servers:
  -
    url: 'http://127.0.0.1:8085'
    description: 'Local environment'
paths:
...
  /api/v1/cities:
    get:
      tags:
        - Cities
      operationId: get_app_city_api_v1_city_list
      parameters:
        -
          name: limit
          in: query
          description: 'Default 500'
          required: false
          content:
            application/json:
              schema:
                type: integer
        -
          name: offset
          in: query
          description: 'Default 0'
          required: false
          content:
            application/json:
              schema:
                type: integer
        -
          name: id
          in: query
          description: 'City uuid'
          required: false
          content:
            application/json:
              schema:
                type: string
        -
          name: title
          in: query
          description: 'City title'
          required: false
          content:
            application/json:
              schema:
                type: string
        -
          name: iata
          in: query
          description: 'City IATA code'
          required: false
          content:
            application/json:
              schema:
                type: string
        -
          name: countryIso2
          in: query
          description: 'Couyntry ISO2 code'
          required: false
          content:
            application/json:
              schema:
                type: string
        -
          name: stateTitle
          in: query
          description: 'State title'
          required: false
          content:
            application/json:
              schema:
                type: string
      responses:
        '200':
          description: 'List of cities.'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CityResponse'
    post:
      tags:
        - Cities
      operationId: post_app_city_api_v1_city_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCityActionRequest'
      responses:
        '200':
          description: 'City created successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityResponse'
  '/api/v1/cities/{id}':
    put:
      tags:
        - Cities
      operationId: put_app_city_api_v1_city_update
      parameters:
        -
          name: id
          in: path
          description: 'City uuid'
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCityActionRequest'
      responses:
        '200':
          description: 'City updated successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityResponse'
    delete:
      tags:
        - Cities
      operationId: delete_app_city_api_v1_city_delete
      parameters:
        -
          name: id
          in: path
          description: 'City uuid'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'City deleted successfully.'
  /api/v1/airports:
    get:
      tags:
        - Airports
      operationId: get_app_airport_api_v1_airport_list
      parameters:
        -
          name: limit
          in: query
          description: 'Default 500'
          required: false
          content:
            application/json:
              schema:
                type: integer
        -
          name: offset
          in: query
          description: 'Default 0'
          required: false
          content:
            application/json:
              schema:
                type: integer
        -
          name: id
          in: query
          description: 'Airprot uuid'
          required: false
          content:
            application/json:
              schema:
                type: string
        -
          name: title
          in: query
          description: 'Airport title'
          required: false
          content:
            application/json:
              schema:
                type: string
        -
          name: iata
          in: query
          description: 'Airport IATA code'
          required: false
          content:
            application/json:
              schema:
                type: string
        -
          name: icao
          in: query
          description: 'Airport ICAO code'
          required: false
          content:
            application/json:
              schema:
                type: string
        -
          name: cityTitle
          in: query
          description: 'City title'
          required: false
          content:
            application/json:
              schema:
                type: string
        -
          name: timezone
          in: query
          description: 'Timezone (e.g. Europe/Oslo)'
          required: false
          content:
            application/json:
              schema:
                type: string
        -
          name: isActive
          in: query
          description: 'Is active (true/false)'
          required: false
          content:
            application/json:
              schema:
                type: boolean
      responses:
        '200':
          description: 'List of airports.'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AirportResponse'
    post:
      tags:
        - Airports
      operationId: post_app_airport_api_v1_airport_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAirportActionRequest'
      responses:
        '200':
          description: 'Airport created successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AirportResponse'
  '/api/v1/airports/{id}':
    put:
      tags:
        - Airports
      operationId: put_app_airport_api_v1_airport_update
      parameters:
        -
          name: id
          in: path
          description: 'Airport uuid'
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAirportActionRequest'
      responses:
        '200':
          description: 'Airport updated successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AirportResponse'
    delete:
      tags:
        - Airports
      operationId: delete_app_airport_api_v1_airport_delete
      parameters:
        -
          name: id
          in: path
          description: 'Airport uuid'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Airport deleted successfully.'
  '/api/v1/airports/{query}':
    get:
      tags:
        - Airports
      operationId: get_app_airport_api_v1_airport_query
      parameters:
        -
          name: query
          in: path
          description: 'Query string to search airports by title, IATA, ICAO, city or country.'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'List of airports.'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QueryChildrenAirportResponse'
components:
  schemas:
...
    RegionResponse:
      required:
        - title
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
      type: object
...
    SubRegionResponse:
      required:
        - title
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        region:
          nullable: true
          oneOf:
            -
              $ref: '#/components/schemas/RegionResponse'
      type: object
...
    TimezoneResponse:
      required:
        - title
        - code
        - utc
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        code:
          type: string
        utc:
          type: string
      type: object
...
    CurrencyResponse:
      required:
        - name
        - code
        - symbol
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
        symbol:
          type: string
      type: object
...
    CountryResponse:
      required:
        - title
        - iso2
        - iso3
        - numericCode
        - phoneCode
        - flag
        - tld
        - longitude
        - latitude
        - timezones
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        nativeTitle:
          type: string
          nullable: true
        iso2:
          type: string
        iso3:
          type: string
        numericCode:
          type: string
        phoneCode:
          type: string
        flag:
          type: string
        tld:
          type: string
        longitude:
          type: number
          format: float
        latitude:
          type: number
          format: float
        altitude:
          type: integer
          nullable: true
        currency:
          nullable: true
          oneOf:
            -
              $ref: '#/components/schemas/CurrencyResponse'
        subRegion:
          nullable: true
          oneOf:
            -
              $ref: '#/components/schemas/SubRegionResponse'
        timezones:
          type: array
          items:
            $ref: '#/components/schemas/TimezoneResponse'
      type: object
...
    StateResponse:
      required:
        - title
        - code
        - latitude
        - longitude
        - country
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        code:
          type: string
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        altitude:
          type: integer
          nullable: true
        type:
          type: string
          nullable: true
        country:
          $ref: '#/components/schemas/CountryResponse'
      type: object
...
    CityResponse:
      required:
        - title
        - longitude
        - latitude
        - country
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        iata:
          type: string
          nullable: true
        longitude:
          type: number
          format: float
        latitude:
          type: number
          format: float
        altitude:
          type: integer
          nullable: true
        state:
          nullable: true
          oneOf:
            -
              $ref: '#/components/schemas/StateResponse'
        country:
          $ref: '#/components/schemas/CountryResponse'
      type: object
    CreateCityActionRequest:
      required:
        - title
        - countryIso2
        - longitude
        - latitude
      properties:
        title:
          type: string
          maxLength: 150
          minLength: 1
        iata:
          type: string
          maxLength: 3
          minLength: 3
          nullable: true
        countryIso2:
          type: string
          maxLength: 2
          minLength: 2
        stateTitle:
          type: string
          maxLength: 150
          nullable: true
        longitude:
          type: number
          format: float
        latitude:
          type: number
          format: float
        altitude:
          type: integer
          nullable: true
      type: object
    UpdateCityActionRequest:
      properties:
        title:
          type: string
          maxLength: 150
          minLength: 1
          nullable: true
        iata:
          type: string
          maxLength: 3
          minLength: 3
          nullable: true
        countryIso2:
          type: string
          maxLength: 2
          minLength: 2
          nullable: true
        stateTitle:
          type: string
          maxLength: 150
          nullable: true
        longitude:
          type: number
          format: float
          nullable: true
        latitude:
          type: number
          format: float
          nullable: true
        altitude:
          type: integer
          nullable: true
      type: object
    AirportResponse:
      required:
        - title
        - iata
        - icao
        - isActive
        - longitude
        - latitude
        - timezone
        - city
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        iata:
          type: string
        icao:
          type: string
        isActive:
          type: boolean
        longitude:
          type: number
          format: float
        latitude:
          type: number
          format: float
        altitude:
          type: integer
          nullable: true
        timezone:
          $ref: '#/components/schemas/TimezoneResponse'
        city:
          $ref: '#/components/schemas/CityResponse'
      type: object
    CreateAirportActionRequest:
      required:
        - title
        - cityTitle
        - iata
        - countryIso2
        - icao
        - timezone
        - longitude
        - latitude
      properties:
        title:
          type: string
          maxLength: 150
          minLength: 1
        cityTitle:
          type: string
          maxLength: 150
          minLength: 1
        iata:
          type: string
          maxLength: 3
          minLength: 3
        countryIso2:
          type: string
          maxLength: 2
          minLength: 2
        icao:
          type: string
          maxLength: 4
          minLength: 4
        timezone:
          type: string
          maxLength: 150
        longitude:
          type: number
          format: float
        latitude:
          type: number
          format: float
        altitude:
          type: integer
          nullable: true
      type: object
    UpdateAirportActionRequest:
      properties:
        icao:
          type: string
          maxLength: 4
          minLength: 4
          nullable: true
        iata:
          type: string
          maxLength: 3
          minLength: 3
          nullable: true
        title:
          type: string
          maxLength: 150
          minLength: 1
          nullable: true
        cityTitle:
          type: string
          maxLength: 150
          minLength: 1
          nullable: true
        countryIso2:
          type: string
          maxLength: 2
          minLength: 2
          nullable: true
        timezone:
          type: string
          maxLength: 150
          minLength: 1
          nullable: true
        isActive:
          type: boolean
          nullable: true
        longitude:
          type: number
          format: float
          nullable: true
        latitude:
          type: number
          format: float
          nullable: true
        altitude:
          type: integer
          nullable: true
      type: object
    QueryChildrenAirportResponse:
      required:
        - title
        - iata
        - country
      properties:
        title:
          type: string
        iata:
          type: string
        country:
          type: string
        children:
          type: array
          items:
            $ref: '#/components/schemas/QueryAirportResponse'
          default: []
      type: object
    QueryAirportResponse:
      required:
        - title
        - iata
        - country
      properties:
        title:
          type: string
        iata:
          type: string
        country:
          type: string
      type: object
  securitySchemes:
    Bearer:
      type: http
      bearerFormat: JWT
      scheme: bearer
security:
  -
    Bearer: []